# infra: speed up unit testing by automatically parallelizing the CPU workers for test runs