--- a/setup.cfg
+++ b/setup.cfg
@@ -4,7 +4,12 @@ test=pytest
 [tool:pytest]
 xfail_strict = true
 addopts =
+    --cov-report term-missing
+    --cov-report html
+    --cov-report xml
+    --cov=braket
     --verbose
+    -n auto
 testpaths = test/unit_tests
 
 [isort]