--- a/setup.py
+++ b/setup.py
@@ -45,7 +45,6 @@ setup(
         "test": [
             "black",
             "botocore",
-            "coverage==5.5",
             "flake8<=5.0.4",
             "isort",
             "jsonschema==3.2.0",