--- a/tox.ini
+++ b/tox.ini
@@ -7,11 +7,7 @@ basepython = python3
 deps =
     {[test-deps]deps}
 commands =
-    coverage run -m pytest {posargs}
-    coverage combine
-    coverage report
-    coverage html
-    coverage xml
+    pytest {posargs}
 extras = test
 
 [testenv:integ-tests]